# BlaBlaNote/docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthUserDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "enum": ["ADMIN", "USER"],
            "type": "string"
          },
          "termsAcceptedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "termsVersion": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role",
          "termsAcceptedAt",
          "termsVersion"
        ],
        "type": "object"
      },
...
      "ErrorResponseDto": {
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          }
        },
        "required": ["statusCode", "message", "error"],
        "type": "object"
      },
...
      "ForgotPasswordDto": {
        "properties": {
          "email": {
            "example": "user@example.com",
            "type": "string"
          }
        },
        "required": ["email"],
        "type": "object"
      },
...
      "LoginDto": {
        "properties": {
          "email": {
            "example": "user@example.com",
            "type": "string"
          },
          "password": {
            "example": "password123",
            "type": "string"
          },
          "rememberMe": {
            "example": true,
            "type": "boolean"
          }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "LoginTokenResponseDto": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "refresh_expires_at",
          "user"
        ],
        "type": "object"
      },
...
      "RefreshTokenResponseDto": {
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": ["access_token", "refresh_token", "refresh_expires_at"],
        "type": "object"
      },
      "RegisterDto": {
        "properties": {
          "email": {
            "example": "user@example.com",
            "type": "string"
          },
          "firstName": {
            "example": "John",
            "type": "string"
          },
          "lastName": {
            "example": "Doe",
            "type": "string"
          },
          "password": {
            "example": "password123",
            "type": "string"
          },
          "termsAccepted": {
            "example": true,
            "type": "boolean"
          },
          "termsVersion": {
            "example": "v1.0",
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "termsAccepted"
        ],
        "type": "object"
      },
...
      "ResetPasswordDto": {
        "properties": {
          "newPassword": {
            "example": "Str0ngPass!word",
            "type": "string"
          },
          "token": {
            "example": "raw_reset_token",
            "type": "string"
          }
        },
        "required": ["token", "newPassword"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "BlaBlaNote SaaS API documentation",
    "title": "BlaBlaNote API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginTokenResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "User login",
        "tags": ["Auth"]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Refresh access and refresh tokens",
        "tags": ["Auth"]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": ["Auth"]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Auth"]
      }
    },
...
  },
  "servers": [],
...
}